4.17个人画像功能完成

+++ src/main/java/com/ngx/boot/vo/portrait/ConsumeData.java
@@ -11,6 +11,6 @@ public class ConsumeData {
 
     private String name;
 
-    private Integer star;
+    private double star;
 
 }
